pyproject.toml
[project]
name = "onchain-fico"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "ipykernel>=6.29.5",
    "ngrok>=1.4.0",
    "numpy<2",
    "packaging>=25.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "scikit-learn>=1.6.1",
    "setuptools>=80.9.0",
    "torch==2.1.2",
    "xgboost>=2.1.4",
]
